fixed some
