+++ module #3. REST API Advanced/core/pom.xml
@@ -80,6 +80,11 @@
             <artifactId>hibernate-java8</artifactId>
             <version>5.4.24.Final</version>
         </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-configuration-processor</artifactId>
+            <optional>true</optional>
+        </dependency>
     </dependencies>
 
     <build>
