+++ module #3. REST API Advanced/core/src/main/java/com/epam/esm/config/DaoConfig.java
@@ -1,0 +1,54 @@
+package com.epam.esm.config;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.hibernate.SessionFactory;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.jdbc.datasource.DriverManagerDataSource;
+import org.springframework.orm.hibernate5.HibernateTransactionManager;
+import org.springframework.orm.hibernate5.LocalSessionFactoryBean;
+
+import javax.sql.DataSource;
+import java.util.Properties;
+
+@Configuration
+@ComponentScan("com.epam.esm")
+@ConfigurationProperties(prefix = "spring.jpa")
+public class DaoConfig {
+    @Getter
+    @Setter
+    private String dialect;
+    @Getter
+    @Setter
+    private String showSql;
+
+    @Bean(name = "dataSource")
+    @ConfigurationProperties(prefix = "spring.datasource")
+    public DataSource getDataSource() {
+        return new DriverManagerDataSource();
+    }
+
+    @Bean(name = "sessionFactory")
+    public SessionFactory getSessionFactory(DataSource dataSource) throws Exception {
+        Properties properties = new Properties();
+        properties.put("hibernate.dialect", dialect);
+        properties.put("hibernate.show_sql", showSql);
+
+        LocalSessionFactoryBean factoryBean = new LocalSessionFactoryBean();
+        factoryBean.setPackagesToScan(new String[] { "com.epam.esm" });
+        factoryBean.setDataSource(dataSource);
+        factoryBean.setHibernateProperties(properties);
+        factoryBean.afterPropertiesSet();
+
+        SessionFactory sessionFactory = factoryBean.getObject();
+        return sessionFactory;
+    }
+
+    @Bean(name = "transactionManager")
+    public HibernateTransactionManager transactionManager(SessionFactory sessionFactory) {
+        return new HibernateTransactionManager(sessionFactory);
+    }
+}
