+++ module #3. REST API Advanced/core/src/main/java/com/epam/esm/entity/GiftCertificate.java
@@ -13,46 +13,40 @@ import java.time.ZonedDateTime;
 import java.util.HashSet;
 import java.util.Set;
 
+@Data
 @NoArgsConstructor
-@EqualsAndHashCode
-@ToString
 @EntityListeners(AuditListener.class)
 @Entity
 @Table(name = GiftCertificateTableConst.TABLE_CERTIFICATE)
 public class GiftCertificate {
-    @Getter
-    @Setter
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.NAME, length = 40)
     private String name;
 
+    @Getter(AccessLevel.NONE)
+    @Setter(AccessLevel.NONE)
     @ManyToMany(fetch = FetchType.EAGER)
     @JoinTable(name = TagToCertificateTableConst.TABLE_TAG_CERT,
             joinColumns = @JoinColumn(name = TagToCertificateTableConst.CERTIFICATE_ID),
     inverseJoinColumns = @JoinColumn(name = TagToCertificateTableConst.TAG_ID))
     private Set<Tag> tagSet = new HashSet<>();
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.DESCRIPTION)
     private String description;
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.PRICE)
     private Double price;
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.CREATE_DATE)
     private ZonedDateTime createDate;
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.LAST_UPDATE_DATE)
     private ZonedDateTime lastUpdateDate;
-    @Getter
-    @Setter
+
     @Column(name = GiftCertificateTableConst.DURATION)
     @Convert(converter = DurationConverter.class)
     private Duration duration;
