+++ module #3. REST API Advanced/core/src/main/resources/application.properties
@@ -1,6 +1,6 @@
-spring.datasource.url=jdbc:mysql://localhost:3306/certificate?serverTimezone=Europe/Minsk
+spring.datasource.url=jdbc:mysql://${DB_HOST}:${DB_PORT}/certificate?serverTimezone=Europe/Minsk
 spring.datasource.username=${DB_USER}
 spring.datasource.password=${DB_PASS}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL57Dialect
-spring.jpa.show-sql=true
+spring.jpa.dialect=org.hibernate.dialect.MySQL57Dialect
+spring.jpa.show_sql=true
