+++ module #3. REST API Advanced/core/src/test/java/com/epam/esm/service/impl/UserServiceImplTest.java
@@ -64,8 +64,8 @@ class UserServiceImplTest {
         QueryCriteria criteria = captor.getValue();
         assertNotNull(criteria.getSortingCriteria());
         assertNotNull(criteria.getSortingOrder());
-        assertEquals(10, criteria.getFirstResult());
-        assertEquals(5, criteria.getResultLimit());
+        assertEquals(20, criteria.getFirstResult());
+        assertEquals(10, criteria.getResultLimit());
 
     }
 }
